Signup Overall Done : Handle Form Data + Create User + Token in Cookie Set
Within this commit, I added the missing part of the signup controller.
So by setting up correctly the frontend as well I'am now able to receive
form data to the backend, and then validate it so as I can create a new
user record from it.

I also added the cookie set, this means that I signed the user data to
create a token from it, and then I stored it in a cookie which will be
available in each new req.

At the end of the process, a simple redirect to /user is performed.