Email Verification Controller
This route handler is the one that will catch the user request for
validating his email address,

As explained in the previous git commit, when the user clicks on the
link that we've sent to him in order to give him the ability to validate
his email address on our platfrom, this is the controller that will
handel that.

How ?
- get from the req.params the user token
- verify if it is valid
- extract the user id from it
- fetch our db
- update the record to have the confirmed_email attribute set to true

Hope that is clear
Massiles.