AuthMiddelware -> ( verifyEmail + verifyLogin Middelwares )
I've decided to break down the prexisting middelware that is meant to
verify that the user is logged in and also that he owns a valid email
address into two sperated middelware that does one thing independently
from the other.

So:
verifyEmail middelware : will verify if the user that is using our
platform currently had validated his email address.

verifyLogin middelware : will verify if the user that is using our
platform currently had logged in.

Why we should do that ?

This is very helpful for creating some sort of engagement & hierarchy , for exemple,
a user that is not registered may have access only on the public pages
like the home, info, about, contact and the reviews pages

Whereas a person that is logged in (so already registered) can have
access not only to that but more  pages as well, like the articals pages,
events pages, but if his email address is not
validated yet, he can not participate to any event or create any
artical, so to speak

it is like filtering, the first filter is the registeration, then the
loggin phase, and finally the email validation. This why you should
always invoke the verifyLogin first, then the verifyEmail after it!

So when a user has a valid email and he is logged in, he has the full
control on the platform.

Besides that, making diffrent middelwares make the code easier to
maintain and to read.

This is my comments on the tweaks, hope that is clear !

Please refer to me if something is wrong !

Thanks
Massiles.