User Redirection Setup
If the user is attempting to validate his email address, so he is now on
his mailbox, and he clicked on the link that we've sent to him, so he is
currently performing an http get request to our backend, and this
controller will handle that.

After handling it, we need to redirect the user, so basicly what I did
is that I check if the user is already loggedin not a long time ago (his
jwt is available in the cookies). If so, we redirect him to the home page. If not,
we redirect him to the login page.

Hope that is clear !
Please refer to me if something is wrong.

Massiles.