List followers params
As the route indicates, we do not need the followerID to list our all
the followers for a given user ( userID ). So our req.params check
should not contain a test for that .

I want to mention, that I dont feel that I will help a lot in the
creation of the friend system, since the used database is a sql one, and
I am only familiar with the noSQL ( mongoDB ) database.

Hope that is clear !
Massiles.

Massiles.