Fixed The System + Optimization
In term of fixes, I tought that `$in : array` operator works like in mongoDB,
but it is not, so I replace it with the ` [Op.in] : array `.

In term of optimization, I used the `Promise.allSettelled` to fire up
multiple async tasks that are not related with each others, to gain time
and performance.

Note: I could of used that shorthand `id : id_array ` instead of ` id :
{ [Op.in] : id_array } `, but I kept the last one for readability
reasons.

_Please refer to me if something went wrong_

Thanks in advance.
Massiles.