Database Cleanup System Overall Done
Within these final tweaks, we are now able not only to purge user
records if confirmed_email is set to false, but to also check if they
are inactiv during the last month.

I created a check that will make sure that we wont delete the user just
if the confirmed_email is set to false and also the createAt attribute
refer to a date before one month ago. All of that by adding a simple if
statement, the simpler the better ( @supernova350  )

Please refer to me if something is wrong.

Thanks.
Massiles.