File Architecture + Implementing the Singleton Paradigm
With this new architecture, we have <controllers> folders and <routes>
folders.

In each <controllers> folder, we have an `index.js` file. In this one, we
import all the controllers within this folder, and export them right
after, so this file will act as our public interface ( singleton paradigm ).

In the <routes> folder, we will import the `index.js` files of each
<controllers> folders and use them to handel the appropriate endpoints.

**Why use this architecture ?**

First of all, this will make our code more organized then just adding
route handler on the main server files, which is already full with other
configuration stuff.

Besides that, this architecture will help us document our code base,
since you can document the controllers and read through the
documentation in the <routes> folder, which is kind of cool. All of that
is made possible by the way that we export and import.

Finnaly, this way of organizing our files is more scalable, since if you
want to add a new endpoint to our platform, all you need to do it to :

1- add one line on the appropriate file inside the <routes> folder.
2- add two lines on the appropriate `index.js` file inside the
<controllers> folder.
3- create your actual route handler in a completelly new file inside the
appropriate <controllers> folder.

That is it. Please refer to me if something is wrong.

Hope that you will like it.
Massiles.