Imported & Setup the express cookie Parser
I've imported the express cookie parser and used it as a middelware so
as we can access the upcomming cookies directly on the
req.cookies.<name-of-the-cookie>. This makes the code more robust and
more reliable then creating our own deconstruction algorithm ( with
split methods ) to get the cookies. ( Reusing code principle ).

Also, with this library setup, our middelwares are more focused about
the task at hand then the preps that it needs. ( Abstraction | Removing
Side effects ).

Hope that is clear.
_Please refer to me if something went wrong._

Thanks in advance.
Massiles.