Fixed Login and Changed Back to Headers
We need web support and we are going to follow internet standards. No major vendors use cookies to store JWT as it can lead to major security issues. Please continue to use Bearer headers to send over the JWT keys. This is safe, this is how we keep our user data secure.